+++ ogre_visualizer/src/ogre_visualizer/displays/particle_cloud_2d_display.cpp
@@ -205,7 +205,7 @@ void ParticleCloud2DDisplay::processMessage()
   clear();
 
   tf::Stamped<tf::Pose> pose( btTransform( btQuaternion( 0.0f, 0.0f, 0.0f ), btVector3( 0.0f, 0.0f, 0.0f ) ),
-                                ros::Time(0ULL), "map" );
+                                ros::Time((uint64_t)0ULL), "map" );
 
   try
   {
