+++ ogre_visualizer/src/ogre_visualizer/displays/planning_display.cpp
@@ -254,7 +254,7 @@ void PlanningDisplay::update( float dt )
 
 void PlanningDisplay::calculateRobotPosition()
 {
-  tf::Stamped<tf::Pose> pose( btTransform( btQuaternion( 0, 0, 0 ), btVector3( 0, 0, 0 ) ), ros::Time(0ULL), displaying_kinematic_path_message_.frame_id );
+  tf::Stamped<tf::Pose> pose( btTransform( btQuaternion( 0, 0, 0 ), btVector3( 0, 0, 0 ) ), ros::Time((uint64_t)0ULL), displaying_kinematic_path_message_.frame_id );
 
   try
   {
