+++ ogre_visualizer/src/ogre_visualizer/helpers/robot.cpp
@@ -616,7 +616,7 @@ void Robot::update( tf::TransformListener* tf, const std::string& target_frame )
       continue;
     }
 
-    tf::Stamped<tf::Pose> pose( btTransform( btQuaternion( 0, 0, 0 ), btVector3( 0, 0, 0 ) ), ros::Time(0ULL), name );
+    tf::Stamped<tf::Pose> pose( btTransform( btQuaternion( 0, 0, 0 ), btVector3( 0, 0, 0 ) ), ros::Time((uint64_t)0ULL), name );
 
     try
     {
